phase-2/backend/pyproject.toml
[tool.poetry]
name = "agentic-todo-backend"
version = "0.1.0"
description = "Backend for AI-Ready Todo App"
authors = ["Developer <developer@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.128.0"
uvicorn = {extras = ["standard"], version = "^0.34.0"}
sqlmodel = "^0.0.31"
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.20"
python-dotenv = "^1.0.1"
asyncpg = "^0.30.0"
alembic = "^1.18.0"
pydantic = "^2.12.0"
pydantic-settings = "^2.12.0"
sqlalchemy = "^2.0.45"
uuid = "^1.30"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
dev = "uvicorn main:app --reload --host 0.0.0.0 --port 8000"
start = "uvicorn main:app --host 0.0.0.0 --port 8000"